+++ test/test.js
@@ -15,6 +15,93 @@ var invalidTypes = [
   'text/plain,wrong',
 ]
 
+describe('typer.format(obj)', function () {
+  it('should format basic type', function () {
+    var str = typer.format({type: 'text', subtype: 'html'})
+    str.should.equal('text/html')
+  })
+
+  it('should format type with suffix', function () {
+    var str = typer.format({type: 'image', subtype: 'svg', suffix: 'xml'})
+    str.should.equal('image/svg+xml')
+  })
+
+  it('should format type with parameter', function () {
+    var str = typer.format({
+      type: 'text',
+      subtype: 'html',
+      parameters: {charset: 'utf-8'}
+    })
+    str.should.equal('text/html; charset=utf-8')
+  })
+
+  it('should format type with parameter that needs quotes', function () {
+    var str = typer.format({
+      type: 'text',
+      subtype: 'html',
+      parameters: {foo: 'bar or "baz"'}
+    })
+    str.should.equal('text/html; foo="bar or \\"baz\\""')
+  })
+
+  it('should format type with parameter with empty value', function () {
+    var str = typer.format({
+      type: 'text',
+      subtype: 'html',
+      parameters: {foo: ''}
+    })
+    str.should.equal('text/html; foo=""')
+  })
+
+  it('should format type with multiple parameters', function () {
+    var str = typer.format({
+      type: 'text',
+      subtype: 'html',
+      parameters: {charset: 'utf-8', foo: 'bar', bar: 'baz'}
+    })
+    str.should.equal('text/html; bar=baz; charset=utf-8; foo=bar')
+  })
+
+  it('should require argument', function () {
+    typer.format.should.throw(/obj.*required/)
+  })
+
+  it('should reject non-objects', function () {
+    typer.format.bind(null, 7).should.throw(/obj.*required/)
+  })
+
+  it('should require type', function () {
+    typer.format.bind(null, {}).should.throw(/invalid type/)
+  })
+
+  it('should reject invalid type', function () {
+    typer.format.bind(null, {type: 'text/'}).should.throw(/invalid type/)
+  })
+
+  it('should require subtype', function () {
+    typer.format.bind(null, {type: 'text'}).should.throw(/invalid subtype/)
+  })
+
+  it('should reject invalid subtype', function () {
+    typer.format.bind(null, {type: 'text', subtype: 'html/'}).should.throw(/invalid subtype/)
+  })
+
+  it('should reject invalid suffix', function () {
+    var obj = {type: 'image', subtype: 'svg', suffix: 'xml\\'}
+    typer.format.bind(null, obj).should.throw(/invalid suffix/)
+  })
+
+  it('should reject invalid parameter name', function () {
+    var obj = {type: 'image', subtype: 'svg', parameters: {'foo/': 'bar'}}
+    typer.format.bind(null, obj).should.throw(/invalid parameter name/)
+  })
+
+  it('should reject invalid parameter value', function () {
+    var obj = {type: 'image', subtype: 'svg', parameters: {'foo': 'bar\u0000'}}
+    typer.format.bind(null, obj).should.throw(/invalid parameter value/)
+  })
+})
+
 describe('typer.parse(string)', function () {
   it('should parse basic type', function () {
     var type = typer.parse('text/html')